+++ src/components/Dropdown/Dropdown.stories.tsx
@@ -1,8 +1,11 @@
+import Button from '@components/Button';
+import Flexbox from '@components-layout/Flexbox';
+import List from '@components-layout/List';
 import { css } from '@emotion/react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { MdAccessibility } from 'react-icons/md';
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { MdPeople } from 'react-icons/md';
 
-import { Dropdown, DropdownCustomItem, DropdownItem } from '.';
+import Dropdown from '.';
 
 export default {
   title: 'Dropdown',
@@ -11,70 +14,70 @@ export default {
     layout: 'fullscreen',
   },
   argTypes: {
-    align: {
-      options: ['left', 'right', 'center'],
-      control: { type: 'radio' },
+    direction: {
+      options: ['left', 'right', 'top', 'bottom'],
+      control: { type: 'select' },
     },
   },
 } as ComponentMeta<typeof Dropdown>;
 
-const DUMMY_TOGGLE = () => {
-  return (
-    <div
-      style={{
-        border: `1px solid #dedede`,
-        borderRadius: '5px',
-        padding: '.5rem',
-      }}
-    >
-      드롭다운토글은커스텀엘레먼트로자유롭게서식을적용할수있습니다테스트를위해아주긴토글을만들었어요
-    </div>
-  );
-};
-
-const DUMMY_CUSTOM_ITEM = css`
-  font-weight: 700;
-  background-color: aqua;
-  padding: 0.5rem;
-`;
-
 const Template: ComponentStory<typeof Dropdown> = (args) => (
-  <Dropdown {...args} toggleElement={<DUMMY_TOGGLE />}>
-    <DropdownItem title="첫 번째 아이템" leftIcon={MdAccessibility} />
-    <DropdownItem title="두 번째 아이템" description="저는 설명이 있는데용" />
-    <DropdownItem
-      title="세 번째 아이템"
-      description="저는 설명과 아이콘도 있어용"
-      leftIcon={MdAccessibility}
-    />
-    <DropdownItem
-      title="세 번째 아이템(무료 alert 포함)"
-      onClick={() => {
-        alert('무료로 제공해드리는 alert입니다');
-      }}
-    />
-    <DropdownItem
-      title="네 번째 아이템"
-      description="저는 오른쪽에 아이콘이 있어용"
-      rightIcon={MdAccessibility}
-    />
-    <DropdownCustomItem css={DUMMY_CUSTOM_ITEM}>
-      저는 커스텀 아이템 이에용
-    </DropdownCustomItem>
-  </Dropdown>
+  <Flexbox
+    css={css`
+      height: 300px;
+      width: 300px;
+      justify-content: center;
+      align-items: center;
+    `}
+  >
+    <Dropdown {...args}>
+      <Dropdown.Trigger>
+        <Button text="팀원 목록" icon={MdPeople} variant="light" />
+      </Dropdown.Trigger>
+      <Dropdown.Menu>
+        <List>
+          <li>김세영</li>
+          <li>백도훈</li>
+          <li>이선민</li>
+          <li>이우재</li>
+          <li>이현빈</li>
+          <li>정주연</li>
+        </List>
+      </Dropdown.Menu>
+    </Dropdown>
+  </Flexbox>
 );
 
-export const Left = Template.bind({});
-Left.args = {
-  align: 'left',
+export const Default = Template.bind({});
+Default.args = {
+  id: 'dropdown1',
+  dropdownLabel: 'Default dropdown',
 };
 
-export const Right = Template.bind({});
-Right.args = {
-  align: 'right',
+export const DirectonTop = Template.bind({});
+DirectonTop.args = {
+  id: 'dropdown2',
+  dropdownLabel: 'Default dropdown',
+  direction: 'top',
 };
 
-export const Center = Template.bind({});
-Center.args = {
-  align: 'center',
+export const DirectonLeft = Template.bind({});
+DirectonLeft.args = {
+  id: 'dropdown3',
+  dropdownLabel: 'Default dropdown',
+  direction: 'left',
+};
+
+export const DirectonRight = Template.bind({});
+DirectonRight.args = {
+  id: 'dropdown4',
+  dropdownLabel: 'Default dropdown',
+  direction: 'right',
+};
+
+export const CollapseOnBlur = Template.bind({});
+CollapseOnBlur.args = {
+  id: 'dropdown5',
+  dropdownLabel: 'Default dropdown',
+  collapseOnBlur: true,
 };
