+++ src/components/Dropdown/index.tsx
@@ -1,166 +1,178 @@
-import Flexbox from '@components/@layout/Flexbox';
+import Container from '@components-layout/Container';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import { DefaultProps } from '@utils/types/DefaultProps';
-import { DefaultPropsWithChildren } from '@utils/types/DefaultPropsWithChildren';
-import { MouseEvent, MouseEventHandler, ReactNode, useRef } from 'react';
-import { IconType } from 'react-icons/lib';
+import { DefaultProps } from '@util-types/DefaultProps';
+import {
+  createContext,
+  Dispatch,
+  MouseEvent as ReactMouseEvent,
+  SetStateAction,
+  useContext,
+  useEffect,
+  useRef,
+  useState,
+} from 'react';
 
-type DropdownAlign = 'left' | 'right' | 'center';
-
-interface DropdownProps extends DefaultPropsWithChildren<HTMLDivElement> {
-  toggleElement: ReactNode;
-  align?: DropdownAlign;
+interface DropdownProps extends DefaultProps<HTMLDivElement> {
+  collapseOnBlur: boolean;
+  dropdownLabel?: string;
+  direction?: 'left' | 'right' | 'top' | 'bottom';
 }
+
+interface IDropdownContext {
+  isOpen?: boolean;
+  setIsOpen?: Dispatch<SetStateAction<boolean>>;
+  collapseOnBlur?: boolean;
+  dropdownLabel?: string;
+  id?: string;
+  direction?: 'left' | 'right' | 'top' | 'bottom';
+  triggerSize?: {
+    width: number;
+    height: number;
+  };
+  setTriggerSize?: Dispatch<
+    SetStateAction<{
+      width: number;
+      height: number;
+    }>
+  >;
+}
+const DropdownContext = createContext<IDropdownContext>({});
 
 const Dropdown = ({
+  collapseOnBlur = false,
+  dropdownLabel = '',
+  id,
+  direction = 'bottom',
   children,
-  toggleElement,
-  align = 'left',
 }: DropdownProps) => {
-  const detailsRef = useRef<HTMLDetailsElement | null>(null);
+  const [isOpen, setIsOpen] = useState<boolean>(false);
+  const [triggerSize, setTriggerSize] = useState<{
+    width: number;
+    height: number;
+  }>({ width: 0, height: 0 });
 
-  const closeDropdown = (e: MouseEvent) => {
-    const { target } = e;
-    if (!target || !detailsRef.current) return;
-    detailsRef.current.open = false;
+  const contextValues = {
+    isOpen,
+    setIsOpen,
+    collapseOnBlur,
+    dropdownLabel,
+    id,
+    direction,
+    triggerSize,
+    setTriggerSize,
   };
 
   return (
-    <DropdownWrapper ref={detailsRef}>
-      <summary>{toggleElement}</summary>
-      <DropdownList onClick={closeDropdown} align={align}>
-        {children}
-      </DropdownList>
-    </DropdownWrapper>
-  );
-};
-
-interface DropdownItemProps extends DefaultProps<HTMLLIElement> {
-  title?: string;
-  leftIcon?: IconType;
-  rightIcon?: IconType;
-  description?: string;
-  onClick?: MouseEventHandler;
-}
-
-const DropdownItem = ({
-  title = '',
-  leftIcon: LeftIcon,
-  rightIcon: RightIcon,
-  description = '',
-  onClick: clickEventHandler,
-}: DropdownItemProps) => {
-  return (
-    <DropdownListItem onClick={clickEventHandler}>
-      {LeftIcon && (
-        <Flexbox alignItems="center">
-          <LeftIcon size={'1.2rem'} />
-        </Flexbox>
-      )}
-      <Flexbox
-        flexDirection="column"
-        gap=".3rem"
-        alignItems="center"
+    <DropdownContext.Provider value={contextValues}>
+      <Container
         css={css`
-          flex: 1;
-          text-align: left;
+          width: auto;
+          height: auto;
+          position: relative;
+          overflow: visible;
         `}
       >
-        {title && <ItemTitle>{title}</ItemTitle>}
-        {description && <ItemDescription>{description}</ItemDescription>}
-      </Flexbox>
-      {RightIcon && (
-        <Flexbox alignItems="center">
-          <RightIcon size={'1.2rem'} />
-        </Flexbox>
-      )}
-    </DropdownListItem>
+        {children}
+      </Container>
+    </DropdownContext.Provider>
   );
 };
 
-const DropdownCustomItem = styled.li<DefaultProps<HTMLLIElement>>(() => {
-  return {
-    cursor: 'pointer',
-  };
-});
+const Trigger = ({ children }: DefaultProps<HTMLDivElement>) => {
+  const { isOpen, setIsOpen, dropdownLabel, id, setTriggerSize } =
+    useContext(DropdownContext);
+  const triggerRef = useRef<HTMLDivElement | null>(null);
 
-export { Dropdown, DropdownItem, DropdownCustomItem };
+  useEffect(() => {
+    if (!triggerRef.current || !setTriggerSize) return;
+    setTriggerSize({
+      width: triggerRef.current.offsetWidth,
+      height: triggerRef.current.offsetHeight,
+    });
+  }, [triggerRef.current]);
 
-const DropdownWrapper = styled.details`
-  width: fit-content;
-  position: relative;
+  return (
+    <>
+      {setIsOpen && (
+        <div
+          onClick={(e: ReactMouseEvent) => {
+            e.stopPropagation();
+            setIsOpen(!isOpen);
+          }}
+          aria-expanded={isOpen}
+          aria-haspopup="true"
+          aria-label={dropdownLabel}
+          id={id}
+          ref={triggerRef}
+        >
+          {children}
+        </div>
+      )}
+    </>
+  );
+};
 
-  & > summary {
-    cursor: pointer;
-    &::marker {
-      content: '';
+const Menu = ({ children }: DefaultProps<HTMLDivElement>) => {
+  const { isOpen, setIsOpen, collapseOnBlur, id, direction, triggerSize } =
+    useContext(DropdownContext);
+  const menuRef = useRef<HTMLDivElement | null>(null);
+
+  const toggleEventHandler = (e: MouseEvent) => {
+    if (!isOpen) return;
+
+    const { target } = e;
+    if (!target || !menuRef.current || !setIsOpen) return;
+
+    if (!menuRef.current.contains(target as Node)) {
+      setIsOpen(false);
     }
-  }
+  };
 
-  &[open] > summary::before {
-    background-color: transparent;
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    z-index: 998;
-    display: block;
-    content: ' ';
-    cursor: default;
-  }
-`;
+  useEffect(() => {
+    if (collapseOnBlur) document.addEventListener('click', toggleEventHandler);
 
-interface ListProps {
-  align: DropdownAlign;
+    return () => {
+      if (collapseOnBlur)
+        document.removeEventListener('click', toggleEventHandler);
+    };
+  }, [isOpen]);
+
+  return (
+    <MenuWrapper
+      ref={menuRef}
+      aria-labelledby={id}
+      direction={direction ?? 'bottom'}
+      triggerSize={triggerSize ?? { width: 0, height: 0 }}
+    >
+      {isOpen && children}
+    </MenuWrapper>
+  );
+};
+
+interface IMenuWrapper {
+  direction: 'left' | 'right' | 'top' | 'bottom';
+  triggerSize: { width: number; height: number };
 }
 
-const DropdownList = styled.ul<ListProps>`
-  ${({ align }) => (align === 'right' ? 'right' : 'left')}: 0;
+const MenuWrapper = styled.div<IMenuWrapper>`
   position: absolute;
-  ${({ align }) =>
-    align === 'center' && 'left: 50%; transform: translateX(-50%);'}
-  z-index: 999;
-  width: fit-content;
-  box-shadow: 0px 1px 3px 1px rgba(0, 0, 0, 0.25);
-  border-radius: 0.2rem;
-
-  & > li:not(:first-of-type) {
-    border-top: 1px solid #ccc;
-  }
-
-  & > li:first-of-type {
-    border-radius: 0.2rem 0.2rem 0 0;
-  }
-
-  & > li:last-of-type {
-    border-radius: 0 0 0.2rem 0.2rem;
-  }
+  ${({ direction, triggerSize }) => {
+    switch (direction) {
+      case 'top':
+        return `bottom: ${triggerSize.height}px; left:0;`;
+      case 'right':
+        return `left: ${triggerSize.width}px; top:0;`;
+      case 'left':
+        return `right: ${triggerSize.width}px; top:0;`;
+      default:
+        return `top: ${triggerSize.height}px; left:0;`;
+    }
+  }}
 `;
 
-const DropdownListItem = styled.li`
-  cursor: pointer;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: white;
-  padding: 0.5rem 0.6rem;
-  gap: 1rem;
+Dropdown.Trigger = Trigger;
+Dropdown.Menu = Menu;
 
-  &:hover {
-    filter: brightness(0.95);
-  }
-`;
-
-const ItemTitle = styled.div`
-  font-size: 1.2rem;
-  font-weight: 700;
-  width: 100%;
-`;
-
-const ItemDescription = styled.div`
-  font-size: 0.8rem;
-  font-weight: 400;
-  width: 100%;
-`;
+export default Dropdown;
+export { DropdownContext };
